Changed behandlingstema for journalføring to correct value.
-Extract common "kodeverk" to separate file.
