+++ client/src/main/kotlin/no/nav/su/se/bakover/client/person/PdlClient.kt
@@ -18,6 +18,7 @@ import no.nav.su.se.bakover.common.objectMapper
 import no.nav.su.se.bakover.domain.AktørId
 import no.nav.su.se.bakover.domain.Fnr
 import no.nav.su.se.bakover.domain.Telefonnummer
+import no.nav.su.se.bakover.domain.Tema
 import no.nav.su.se.bakover.domain.person.KunneIkkeHentePerson
 import no.nav.su.se.bakover.domain.person.KunneIkkeHentePerson.FantIkkePerson
 import no.nav.su.se.bakover.domain.person.KunneIkkeHentePerson.IkkeTilgangTilPerson
@@ -25,10 +26,6 @@ import no.nav.su.se.bakover.domain.person.KunneIkkeHentePerson.Ukjent
 import org.slf4j.LoggerFactory
 import org.slf4j.MDC
 import java.time.LocalDate
-
-const val NAV_CONSUMER_TOKEN = "Nav-Consumer-Token"
-const val NAV_TEMA = "Tema"
-const val SUP = "SUP"
 
 internal class PdlClient(
     private val pdlUrl: String,
@@ -98,8 +95,8 @@ internal class PdlClient(
         val token = tokenOppslag.token()
         val (_, response, result) = "$pdlUrl/graphql".httpPost()
             .header("Authorization", MDC.get("Authorization"))
-            .header(NAV_CONSUMER_TOKEN, "Bearer $token")
-            .header(NAV_TEMA, SUP)
+            .header("Nav-Consumer-Token", "Bearer $token")
+            .header("Tema", Tema.SUPPLERENDE_STØNAD.value)
             .header("Accept", "application/json")
             .header("Content-Type", "application/json")
             .body(objectMapper.writeValueAsString(pdlRequest))
