+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/oppgave/OppgaveConfig.kt
@@ -1,13 +1,13 @@
 package no.nav.su.se.bakover.domain.oppgave
 
 import no.nav.su.se.bakover.domain.AktørId
+import no.nav.su.se.bakover.domain.Behandlingstema
+import no.nav.su.se.bakover.domain.Behandlingstype
 import no.nav.su.se.bakover.domain.NavIdentBruker
+import no.nav.su.se.bakover.domain.Oppgavetype
 import no.nav.su.se.bakover.domain.journal.JournalpostId
 import java.util.UUID
 
-/**
- * https://jira.adeo.no/browse/OH-580
- */
 sealed class OppgaveConfig {
     abstract val journalpostId: JournalpostId?
     abstract val søknadId: UUID
@@ -36,39 +36,6 @@ sealed class OppgaveConfig {
         override val journalpostId: JournalpostId? = null
         override val behandlingstema = Behandlingstema.SU_UFØRE_FLYKNING
         override val behandlingstype = Behandlingstype.FØRSTEGANGSSØKNAD
-        override val oppgavetype = Oppgavetype.TIL_ATTESTERING
-    }
-
-    /**
-     * https://kodeverk-web.nais.adeo.no/kodeverksoversikt/kodeverk/Behandlingstyper
-     * https://github.com/navikt/kodeverksmapper/blob/master/web/src/main/resources/underkategori.csv
-     */
-    enum class Behandlingstype(val value: String) {
-        FØRSTEGANGSSØKNAD("ae0245"),
-        BEHANDLE_VEDTAK("ae0004");
-
-        override fun toString() = this.value
-    }
-
-    /**
-     * https://kodeverk-web.nais.adeo.no/kodeverksoversikt/kodeverk/Behandlingstema
-     * https://github.com/navikt/kodeverksmapper/blob/master/web/src/main/resources/underkategori.csv
-     */
-    enum class Behandlingstema(val value: String) {
-        SU_UFØRE_FLYKNING("ab0431"),
-        SU_ALDER("ab0432");
-
-        override fun toString() = this.value
-    }
-
-    /**
-     * https://kodeverk-web.nais.adeo.no/kodeverksoversikt/kodeverk/Oppgavetyper
-     * https://github.com/navikt/kodeverksmapper/blob/master/web/src/main/resources/oppgavetype.csv
-     */
-    enum class Oppgavetype(val value: String) {
-        BEHANDLE_SAK("BEH_SAK"),
-        TIL_ATTESTERING("ATT");
-
-        override fun toString() = this.value
+        override val oppgavetype = Oppgavetype.ATTESTERING
     }
 }
